+++ CurrencyConverter/Public/Get-ExchangeRate.ps1
@@ -1,0 +1,45 @@
+function Get-ExchangeRate {
+    <#
+    .SYNOPSIS
+        Returns the exchange rates for a specified currency
+
+    .DESCRIPTION
+        Uses the Open Exchange Rate API (https://www.exchangerate-api.com/docs/free) to retrieve currency rates.
+        Caches the conversion rates for the specified currency (valid for 24 hours).
+
+    .PARAMETER Currency
+        The currency to return rates for.
+
+    .EXAMPLE
+        Get-ExchangeRate -Currency USD
+
+        Description
+        -----------
+        Returns the exchange rates for US Dollar as a PowerShell object.
+    #>
+    param(
+        [validateset('AED', 'AFN', 'ALL', 'AMD', 'ANG', 'AOA', 'ARS', 'AUD', 'AWG', 'AZN', 'BAM', 'BBD', 'BDT', 'BGN', 'BHD', 'BIF', 'BMD', 'BND', 'BOB', 'BRL', 'BSD', 'BTN', 'BWP', 'BYN', 'BZD', 'CAD', 'CDF', 'CHF', 'CLP', 'CNY', 'COP', 'CRC', 'CUP', 'CVE', 'CZK', 'DJF', 'DKK', 'DOP', 'DZD', 'EGP', 'ERN', 'ETB', 'EUR', 'FJD', 'FKP', 'FOK', 'GBP', 'GEL', 'GGP', 'GHS', 'GIP', 'GMD', 'GNF', 'GTQ', 'GYD', 'HKD', 'HNL', 'HRK', 'HTG', 'HUF', 'IDR', 'ILS', 'IMP', 'INR', 'IQD', 'IRR', 'ISK', 'JEP', 'JMD', 'JOD', 'JPY', 'KES', 'KGS', 'KHR', 'KID', 'KMF', 'KRW', 'KWD', 'KYD', 'KZT', 'LAK', 'LBP', 'LKR', 'LRD', 'LSL', 'LYD', 'MAD', 'MDL', 'MGA', 'MKD', 'MMK', 'MNT', 'MOP', 'MRU', 'MUR', 'MVR', 'MWK', 'MXN', 'MYR', 'MZN', 'NAD', 'NGN', 'NIO', 'NOK', 'NPR', 'NZD', 'OMR', 'PAB', 'PEN', 'PGK', 'PHP', 'PKR', 'PLN', 'PYG', 'QAR', 'RON', 'RSD', 'RUB', 'RWF', 'SAR', 'SBD', 'SCR', 'SDG', 'SEK', 'SGD', 'SHP', 'SLE', 'SOS', 'SRD', 'SSP', 'STN', 'SYP', 'SZL', 'THB', 'TJS', 'TMT', 'TND', 'TOP', 'TRY', 'TTD', 'TVD', 'TWD', 'TZS', 'UAH', 'UGX', 'USD', 'UYU', 'UZS', 'VES', 'VND', 'VUV', 'WST', 'XAF', 'XCD', 'XDR', 'XOF', 'XPF', 'YER', 'ZAR', 'ZMW', 'ZWL')]
+        [parameter(Mandatory)]
+        [string]
+        $Currency
+    )
+
+    $FromFile = "$PSScriptRoot\${Currency}.json"
+    $CachedCurrency = Test-Path $FromFile
+
+    if ($CachedCurrency) {
+        $ExchangeRates = Get-Content $FromFile | ConvertFrom-Json
+    }
+
+    if (-not $CachedCurrency -or ((Get-Date) -gt ($ExchangeRates.time_next_update_unix | ConvertFrom-UnixTime))) {
+        $ExchangeRates = Invoke-RestMethod "https://open.er-api.com/v6/latest/${Currency}"
+        $ExchangeRates | ConvertTo-Json | Out-File "$PSScriptRoot\${Currency}.json" -Force
+    }
+
+    if ($ExchangeRates) {
+        return $ExchangeRates
+    }
+    else {
+        Write-Error "Failed to retrieve rates for $Currency."
+    }
+}