+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## !Deploy
+
+* Implemented `Get-ExchangeRate` to return the full result from the API as a PowerShell object.
+
 ## [0.0.1] - 2024-02-07
 
 * Implemented `Convert-Currency` to perform currency conversions via the ExchangeRate-API Open Access API.